lab5.1/requirements.txt
Flask==2.2.2
tensorflow==2.11.0
numpy==1.23.5
Pillow==9.2.0